backend/requirements.txt
# backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3.1
PyJWT>=2.8
celery>=5.3
redis>=5.0
langchain
langchain-community
python-dotenv
unstructured[pdf]
tiktoken
langchain-google-genai
google-generativeai
pandas
requests
watchdog
scikit-learn
# The 'openai' line has been removed

# MCP Server Dependencies
mcp>=1.0.0
asyncio-mqtt>=0.16.0